Move remote_config.dart to resources/
